+++ examples/index.html
@@ -6,6 +6,7 @@
 
 <body>
     <script src='./index.js'></script>
+    <canvas id="triangle"></canvas>
 </body>
 
 </html>
