+++ examples/triangle/main.js
@@ -1,32 +1,35 @@
-import { Matter, Shader, ShaderKind, Program, Context, BufferKind } from "../../wasm/warp_engine";
-import { booted } from "../../wasm/warp_engine_bg";
+const warp = import('../../wasm/warp_engine');
 
-function main() {
+//import { Matter, Shader, ShaderKind, Program, Context, BufferKind } from "../../wasm/warp_engine";
+//import { booted } from "../../wasm/warp_engine_bg";
+
+function main(module) {
     let size = [800, 600];
     //let config = AppConfig::new ("Test", size);
     //let mut app = App::new (config);
 
-    let vertices = Float32Array([-0.5, 0.5, 0.0, -0.5, -0.5, 0.0, 0.5, -0.5, 0.0]);
+    let vertices = new Float32Array([-0.5, 0.5, 0.0, -0.5, -0.5, 0.0, 0.5, -0.5, 0.0]);
 
-    let indices = Uint16Array([0, 1, 2]);
+    let indices = new Uint16Array([0, 1, 2]);
     let count = 3;
 
-    let context = Context.new("canvas");
-    let matter = Matter.new(context, vertices, indices);
+    //let context = Context.new("canvas");
+    let context = document.getElementById("triangle").getContext("webgl2");
+    let matter = module.Matter.new(context, vertices, indices);
 
     /*================ Shaders ====================*/
 
     // Vertex shader source code
     let vert_code = "attribute vec3 coordinates;void main(void){gl_Position = vec4(coordinates, 1.0);}";
 
-    let vert_shader = Shader.new(context, vert_code, ShaderKind.Vertex);
+    let vert_shader = module.Shader.new(context, vert_code, ShaderKind.Vertex);
 
     //fragment shader source code
     let frag_code = "void main(void){gl_FragColor = vec4(1, 0.5, 0.0, 1);}";
     // Create fragment shader object
-    let frag_shader = Shader.new(context, frag_code, ShaderKind.Fragment);
+    let frag_shader = module.Shader.new(context, frag_code, ShaderKind.Fragment);
 
-    let shader_program = Program.new(context, vert_shader, frag_shader);
+    let shader_program = module.Program.new(context, vert_shader, frag_shader);
 
     // Use the combined shader program object
     shader_program.use();
@@ -78,4 +81,10 @@ function main() {
     }
 
 }
-booted.then(main);
+
+
+function load() {
+    warp.then(module => main(module))
+}
+
+window.onload = load;
