+++ glenum-bindgen/Cargo.toml
@@ -4,6 +4,5 @@ version = "0.1.0"
 authors = ["Theo Dedeken <theo.dedeken@telenet.be>"]
 
 [dependencies]
-wasm-bindgen = "0.2.11"
-serde = "1.0"
-serde_derive = "1.0"
+wasm-bindgen = { git = "https://github.com/rustwasm/wasm-bindgen" }
+#wasm-bindgen = "0.2.11"
