+++ warp-engine/Cargo.toml
@@ -7,6 +7,7 @@ authors = ["Theo Dedeken <theo.dedeken@telenet.be>"]
 crate-type = ["cdylib"]
 
 [dependencies]
-wasm-bindgen = "0.2.11"
+wasm-bindgen = { git = "https://github.com/rustwasm/wasm-bindgen" }
+#wasm-bindgen = "0.2.11"
 glenum-bindgen = { path = "../glenum-bindgen" }
 webgl2-bindgen = { path = "../webgl2-bindgen" }
