+++ webgl2-bindgen/Cargo.toml
@@ -4,5 +4,6 @@ version = "0.1.0"
 authors = ["Theo Dedeken <theo.dedeken@telenet.be>"]
 
 [dependencies]
-wasm-bindgen = "0.2.11"
+wasm-bindgen = { git = "https://github.com/rustwasm/wasm-bindgen" }
+#wasm-bindgen = "0.2.11"
 glenum-bindgen = { path = "../glenum-bindgen" }
