+++ webgl2-bindgen/src/lib.rs
@@ -1,19 +1,22 @@
+//! COPY of file in crate webgl2_bindgen, currently necessary to make the bindings work.
+//!
 //! Bindings for all objects and method associated with WebGL2
 //!
 //! Documentation taken straight from https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext
 //! and https://developer.mozilla.org/en-US/docs/Web/API/WebGL2RenderingContext
-#![feature(proc_macro, wasm_custom_section, wasm_import_module)]
+#![feature(use_extern_macros, wasm_import_module)]
 
+//TODO: all types to refs (or derive copy?)
 //TODO: arraybufferview as enum and impl method to map to different functions
 //TODO: all pub types in function to references
 //TODO: possible solution of differen getParameter methods is to add accessors on different types
-
 extern crate glenum_bindgen;
 extern crate wasm_bindgen;
 
 use glenum_bindgen::*;
 use wasm_bindgen::prelude::*;
 
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type HTMLDocument;
@@ -60,10 +63,10 @@ impl WebGL2RenderingContext {
 
 /// WebGL2RenderingContext
 #[wasm_bindgen]
+#[derive(Clone, Copy)]
 extern "C" {
     /// The WebGL2RenderingContext interface provides the OpenGL ES 3.0 rendering context
     /// for the drawing surface of an HTML <canvas> element.
-    #[derive(Copy, Clone)]
     pub type WebGL2RenderingContext;
 
     /// The `WebGLRenderingContext.canvas` property is a read-only reference to the `HTMLCanvasElement`
@@ -90,8 +93,10 @@ extern "C" {
 
     /// The `WebGLRenderingContext.getContextAttributes()` method returns a `WebGLContextAttributes`
     /// object that contains the actual context parameters. Might return `null`, if the context is lost.
+    /* FIXME: the object is not defined when imported
     #[wasm_bindgen(method, js_name = getContextAttributes)]
     pub fn get_context_attributes(this: &WebGL2RenderingContext) -> WebGLContextAttributes;
+    */
 
     /// The WebGLRenderingContext.isContextLost() method returns a Boolean indicating whether or not
     /// the WebGL context has been lost.
@@ -331,7 +336,7 @@ extern "C" {
 
     /// The `WebGLRenderingContext.bindBuffer()` method of the WebGL API binds a given WebGLBuffer to a target.
     #[wasm_bindgen(method, js_name = bindBuffer)]
-    pub fn bind_buffer(this: &WebGL2RenderingContext, target: BufferKind, buffer: WebGLBuffer);
+    pub fn bind_buffer(this: &WebGL2RenderingContext, target: BufferKind, buffer: &WebGLBuffer);
 
     /// TODO maybe add a method for every buffer type
 
@@ -343,9 +348,9 @@ extern "C" {
         target: BufferKind,
         srcData: Vec<u8>,
         usage: DataHint,
-        srcOffset: u32,
-        length: u32,
     );
+
+    //TODO buffer_data with offset
 
     /// The `WebGLRenderingContext.bufferSubData()` method of the WebGL API updates a subset of a
     /// buffer object's data store.
@@ -609,7 +614,11 @@ extern "C" {
     /// The `WebGLRenderingContext.attachShader()`  method of the WebGL API attaches either a fragment or
     /// vertex WebGLShader to a WebGLProgram.
     #[wasm_bindgen(method, js_name = attachShader)]
-    pub fn attach_shader(this: &WebGL2RenderingContext, program: WebGLProgram, shader: WebGLShader);
+    pub fn attach_shader(
+        this: &WebGL2RenderingContext,
+        program: &WebGLProgram,
+        shader: &WebGLShader,
+    );
 
     /// The `WebGLRenderingContext.bindAttribLocation()` method of the WebGL API binds a generic vertex index
     /// to an attribute variable.
@@ -624,7 +633,7 @@ extern "C" {
     /// The `WebGLRenderingContext.compileShader()` method of the WebGL API compiles a GLSL shader into binary
     /// data so that it can be used by a WebGLProgram.
     #[wasm_bindgen(method, js_name = compileShader)]
-    pub fn compile_shader(this: &WebGL2RenderingContext, shader: WebGLShader);
+    pub fn compile_shader(this: &WebGL2RenderingContext, shader: &WebGLShader);
 
     /// The `WebGLRenderingContext.createProgram()` method of the WebGL API creates and initializes a WebGLProgram object.
     #[wasm_bindgen(method, js_name = createProgram)]
@@ -633,7 +642,7 @@ extern "C" {
     /// The `WebGLRenderingContext.createShader()` method of the WebGL API creates a WebGLShader that can then be configured
     /// further using `WebGLRenderingContext.shaderSource()` and `WebGLRenderingContext.compileShader()`.
     #[wasm_bindgen(method, js_name = createShader)]
-    pub fn create_shader(this: &WebGL2RenderingContext) -> WebGLShader;
+    pub fn create_shader(this: &WebGL2RenderingContext, kind: ShaderKind) -> WebGLShader;
 
     /// The `WebGLRenderingContext.deleteProgram()` method of the WebGL API deletes a given WebGLProgram object. This method
     /// has no effect if the program has already been deleted.
@@ -687,7 +696,7 @@ extern "C" {
 
     /// The `WebGLRenderingContext.getShaderSource()` method of the WebGL API returns the source code of a WebGLShader as a DOMString.
     #[wasm_bindgen(method, js_name = getShaderSource)]
-    pub fn get_shader_source(this: &WebGL2RenderingContext, shader: WebGLShader) -> String;
+    pub fn get_shader_source(this: &WebGL2RenderingContext, shader: &WebGLShader) -> String;
 
     /// The `WebGLRenderingContext.isProgram()` method of the WebGL API returns true if the passed WebGLProgram is valid, false otherwise.
     #[wasm_bindgen(method, js_name = isProgram)]
@@ -699,11 +708,11 @@ extern "C" {
 
     /// The `WebGLRenderingContext.linkProgram()` method of the WebGL API links a given WebGLProgram to the attached vertex and fragment shaders.
     #[wasm_bindgen(method, js_name = linkProgram)]
-    pub fn link_program(this: &WebGL2RenderingContext, program: WebGLProgram);
+    pub fn link_program(this: &WebGL2RenderingContext, program: &WebGLProgram);
 
     /// The `WebGLRenderingContext.shaderSource()` method of the WebGL API sets the source code of a WebGLShader.
     #[wasm_bindgen(method, js_name = shaderSource)]
-    pub fn shader_source(this: &WebGL2RenderingContext, shader: WebGLShader, source: &str);
+    pub fn shader_source(this: &WebGL2RenderingContext, shader: &WebGLShader, source: &str);
 
     /// The `WebGLRenderingContext.useProgram()` method of the WebGL API sets the specified WebGLProgram as part of the current rendering state.
     #[wasm_bindgen(method, js_name = useProgram)]
@@ -1771,6 +1780,8 @@ extern "C" {
 }
 
 /// WebGLContextAttributes
+/* FIXME: not found when exported
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLContextAttributes;
@@ -1791,44 +1802,52 @@ extern "C" {
     #[wasm_bindgen(method, getter = powerPreference)]
     pub fn power_preference(this: &WebGLContextAttributes) -> String;
 }
+*/
 
 /// WebGLBuffer
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLBuffer;
 }
 
 /// WebGLFramebuffer
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLFramebuffer;
 }
 
 /// WebGLRenderbuffer
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLRenderbuffer;
 }
 
 /// WebGLTexture
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLTexture;
 }
 
 /// WebGLProgram
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLProgram;
 }
 
 /// WebGLShader
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLShader;
 }
 
 /// WebGLShaderPrecisionFormat;
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLShaderPrecisionFormat;
@@ -1847,6 +1866,7 @@ extern "C" {
 }
 
 /// WebGLActiveInfo
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLActiveInfo;
@@ -1865,36 +1885,42 @@ extern "C" {
 }
 
 /// WebGLUniformLocation
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLUniformLocation;
 }
 
 /// WebGLQuery
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLQuery;
 }
 
 /// WebGLSampler
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLSampler;
 }
 
 /// WebGLSync
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLSync;
 }
 
 /// WebGLTransformFeedback
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLTransformFeedback;
 }
 
 /// WebGLVertexArrayObject
+#[derive(Clone, Copy)]
 #[wasm_bindgen]
 extern "C" {
     pub type WebGLVertexArrayObject;
